Adicionado tipo de chave integer

+++ br.com.project.checkskills/src/main/java/br/com/project/checkskills/entities/dadosbasicos/CargoEntity.java
@@ -2,6 +2,6 @@ package br.com.project.checkskills.entities.dadosbasicos;
 
 import br.com.project.checkskills.utils.BaseEntity;
 
-public class CargoEntity extends BaseEntity<Long> {
+public class CargoEntity extends BaseEntity<Integer> {
 
 }
